FEAT: started on refactoring of TDFSpectrumReader
